Limit support to Windows 10/11
To optimise this file (and to make it easier for me to edit!) - I have dropped the configurations for unsupported versions of Windows (anything below Windows 10 20H1 or Windows 11 21H2).

Additionally, I will be adding configs for Windows 10 20H1+ on a "regular" basis. Windows 11 configs will be updated when/if I have a need for them (if you do need these generated, please send a copy of the termsrv.dll so I can generate this for you and add it to the repo).